requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.2
django-allauth==0.35.0
django-rest-auth==0.9.3
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
idna==2.6
oauthlib==2.0.6
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
